express-validator and corrections

+++ src/router/admin/products.js
@@ -3,6 +3,7 @@ const { Router } = require('express');
 const router = Router();
 const upload = require("./multer");
 const cloudinary = require("./cloudinary");
+const { validationResult } = require("express-validator");
 
 const fs = require("fs");
 const dotenv = require("dotenv");
@@ -11,6 +12,7 @@ dotenv.config();
 
 // imports
 const Product = require("../../models/Product");
+const { productsValidate, productsUpdateValidate } = require("../../helper/validator.helper");
 
 // adding multiple products images
 router.post('/upload-multiple', upload.array('image'), async (req, res) => {
@@ -58,42 +60,75 @@ router.post('/upload-single', upload.single('image'), async (req, res) => {
 
 
 // create product with single image
-router.post('/', upload.single('productImage'), async (req, res) => {
+router.post('/', upload.single('image'), productsValidate(), async (req, res) => {
+    try {
 
-    const { productName, productDesc,
-        productCost, productQuantity, productStatus, categoryId, categoryName } = req.body;
-
-    const findProduct = await Product.findOne({ productName })
-    if (findProduct) {
-        return res.status(400).json({ Message: 'Product already exists!!!' })
-    }
-
-    else {
-        const uploader = async (path) => {
-            return await cloudinary.uploads(path, 'Image');
+        let errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            console.log(errors.array());
+            return res.json({ errors: errors.array() })
         }
 
-        const url = []
-        const file = req.file;
-        const { path } = file;
-        const uploadedImage = await uploader(path)
-        url.push(uploadedImage)
-        fs.unlinkSync(path)
+        const { productName, productDesc,
+            productCost, productQuantity, productStatus, categoryId, categoryName } = req.body;
 
-        // const imgUrl = url.map((data) =>(data.url))
-        // console.log(url[0].url)
+        const findProduct = await Product.findOne({ productName })
+        if (findProduct) {
+            return res.status(400).json({ Message: 'Product already exists!!!' })
+        }
 
-        const product = await Product.create({
-            productName, productDesc, productImage: url[0].url,
-            productCost, productQuantity, productStatus, categoryId, categoryName
-        })
-        return res.status(201).json(product)
+        else {
+            const uploader = async (path) => {
+                return await cloudinary.uploads(path, 'Image');
+            }
+
+            const url = []
+            const file = req.file;
+            console.log(file.mimetype)
+
+            const { path } = file;
+            // console.log(path)
+            if (file.mimetype !== 'image/jpeg' || file.mimetype !== 'image/jpg' || file.mimetype !== 'image/png' || file.mimetype !== '') {
+                return res.status(422).json({ Message: 'File not supported!!!' })
+            }
+
+            if (file.mimetype === 'image/jpeg' || file.mimetype === 'image/jpg' || file.mimetype === 'image/png') {
+                const uploadedImage = await uploader(path)
+                url.push(uploadedImage)
+                fs.unlinkSync(path)
+
+                // const imgUrl = url.map((data) =>(data.url))
+                // console.log(url[0].url)
+
+                const product = await Product.create({
+                    productName, productDesc, productImage: url[0].url,
+                    productCost, productQuantity, productStatus, categoryId, categoryName
+                })
+                console.log(product)
+                return res.status(201).json(product)
+            }
+        }
+    } catch (error) {
+        console.log(error)
     }
 
 });
 
 // create product with multiple images
-router.post('/product-images', upload.array('productImage'), async (req, res) => {
+router.post('/product-images', upload.array('image'), productsValidate(), async (req, res) => {
+
+    let errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        console.log(errors.array());
+        return res.json({ errors: errors.array() })
+    }
+    const img_array = ['image/jpeg', 'image/png', 'image/jpg']
+
+    for (const file of req.files) {
+        if (!img_array.includes(file.mimetype)) {
+            return res.status(422).json({ Message: 'File format not supported!!!' })
+        }
+    }
 
     const { productName, productDesc,
         productCost, productQuantity, productStatus, categoryId, categoryName } = req.body;
@@ -111,6 +146,7 @@ router.post('/product-images', upload.array('productImage'), async (req, res) =>
         var urls = []
         const files = req.files;
         for (const file of files) {
+            // console.log(file)
             const { path } = file;
             const newPath = await uploader(path)
             urls.push(newPath.url)
@@ -118,13 +154,12 @@ router.post('/product-images', upload.array('productImage'), async (req, res) =>
         }
 
         // urls.map((data) =>{console.log(data)})
-            Product.create({
-                productName, productDesc, productImage: urls,
-                productCost, productQuantity, productStatus, categoryId, categoryName
-            })
-        return res.status(201).json({Message: 'Products created successfully!!!'})
+        Product.create({
+            productName, productDesc, productImage: urls,
+            productCost, productQuantity, productStatus, categoryId, categoryName
+        })
+        return res.status(201).json({ Message: 'Products created successfully!!!' })
     }
-
 });
 
 // read/list all products
@@ -157,7 +192,13 @@ router.get('/:id', async (req, res) => {
 });
 
 // find and update a product by id
-router.patch('/:id', async (req, res) => {
+router.patch('/:id', productsUpdateValidate(), async (req, res) => {
+
+    let errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        console.log(errors.array());
+        return res.json({ errors: errors.array() })
+    }
 
     const { id } = req.params;
